Limit visibility of instance variables in node specs
Instance variables containing AST nodes for node specs are defined in a
"before" hook inside a RSpec.configure block. This is because there is
no "describe" wrapping all the node specs in which the "before" hook
could be placed instead.

Because of this, before this commit the instance variables were visible
for all specs (with a possibility of conflicts). This commit uses RSpec
metadata to limit their visibility.